% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.R
\name{calc}
\alias{calc}
\title{Calculate and go through the quadrapole data chemical analysis}
\usage{
calc(dat, weights)
}
\arguments{
\item{dat}{is the .csv file made from the read_rep()}

\item{weights}{is the .csv file with all the chemical weights}
}
\description{
Calculate and go through the quadrapole data chemical analysis
}
