% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelFiles.R
\name{excelFiles}
\alias{excelFiles}
\title{Create excel file for the quadrapole calculations and data}
\usage{
excelFiles(dat = calculations, mypath = NULL, rewrite = TRUE)
}
\arguments{
\item{dat}{the calculations outputted from the `calc()` function}

\item{mypath}{will determine where we save the excel file. Entered as a string.}

\item{rewrite}{}
}
\description{
Create excel file for the quadrapole calculations and data
}
