% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Rep.R
\name{read_rep}
\alias{read_rep}
\title{Read in rep files from quadrapole machine}
\usage{
read_rep(file = "RawData/casting pins_102301-303_u_9-5-18.rep")
}
\arguments{
\item{file}{= path to the .rep file you need. Entered as a string.}
}
\description{
Read in rep files from quadrapole machine
}
