% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results.R
\name{results}
\alias{results}
\title{Combines the read_rep, calc, and excelfiles Functions.}
\usage{
results(file_in = NULL, weights = NULL, file_out = NULL)
}
\arguments{
\item{file_in}{= path to the .rep file you need. Entered as a string.}

\item{weights}{= a path to the .csv file with the needed weights. Entered as a string.}

\item{file_out}{= a string name of what you want the function to output the file as. Entered as a string.}
}
\description{
Combines the read_rep, calc, and excelfiles Functions.
}
